+++ NEWS.md
@@ -1,3 +1,7 @@
-# 0.1.1
+# shiny.benchmark (development)
+
+Create reports automatically using `report_file` argument in `benchmark` function.
+
+# shiny.benchmark 0.1.1
 
 Adding a minimal example using `shinytest2`